+++ src/dialogs/tools/dialogcutspline.h
@@ -74,20 +74,20 @@ public:
      * @brief getSplineId return id base point of line
      * @return id
      */
-    inline qint64     getSplineId() const {return splineId;}
+    inline quint32     getSplineId() const {return splineId;}
     /**
      * @brief setSplineId set id spline
      * @param value id
      * @param id don't show this id in list
      */
-    void              setSplineId(const qint64 &value, const qint64 &id);
+    void              setSplineId(const quint32 &value, const quint32 &id);
 public slots:
     /**
      * @brief ChoosedObject gets id and type of selected object. Save right data and ignore wrong.
      * @param id id of point or detail
      * @param type type of object
      */
-    virtual void      ChoosedObject(qint64 id, const Scene::Scenes &type);
+    virtual void      ChoosedObject(quint32 id, const Scene::Scenes &type);
     /**
      * @brief DialogAccepted save data and emit signal about closed dialog.
      */
@@ -109,7 +109,7 @@ private:
     /**
      * @brief splineId keep id of spline
      */
-    qint64            splineId;
+    quint32            splineId;
 };
 
 #endif // DIALOGCUTSPLINE_H
